requirements.txt
alembic==1.8.1
bcrypt==4.0.1
blinker==1.4
certifi==2016.2.28
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
flask-sock==0.5.2
Flask-SQLAlchemy==3.0.2
Turbo-Flask==0.8.3.dev0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.3.0
pycparser==2.18
six==1.11.0
SQLAlchemy==1.2.7
Werkzeug==2.2.2
WTForms==2.1